backport move status/sort constants
