+++ twitcher/wps_restapi/processes/processes.py
@@ -6,13 +6,11 @@ from six.moves.urllib.error import URLError
 from time import sleep
 from twitcher.adapter import servicestore_factory, jobstore_factory
 from twitcher.datatype import Job as JobDB
-from twitcher.exceptions import JobRegistrationError
-from twitcher.owsexceptions import OWSNoApplicableCode
-from twitcher.utils import get_any_id, raise_on_xml_exception
+from twitcher.utils import get_any_id
 from twitcher.wps_restapi import swagger_definitions as sd
 from twitcher.wps_restapi.utils import *
 from twitcher.wps_restapi.jobs.jobs import check_status
-from twitcher.wps_restapi import status
+from twitcher import status
 from owslib.wps import WebProcessingService, WPSException, ComplexDataInput, is_reference
 from owslib.util import clean_ows_url
 from lxml import etree
